% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfacta.R
\name{myfacta}
\alias{myfacta}
\title{myfacta}
\usage{
myfacta(S = NULL, R = NULL, m = NULL)
}
\arguments{
\item{S}{input sample var-cov matrix}

\item{R}{input sample correlation matrix}

\item{m}{input number of factors}
}
\value{
a heat map of the matrix R, proportion of total variance due to the jth factor, the matrix Psi,
        the residual matrix, and the communalities
}
\description{
This function generates a heat map of the matrix R, proportion of total variance due to the jth factor, the matrix Psi,
    the residual matrix, and the communalities.
}
\examples{
S = matrix(c(1.0, 0.63, 0.45, 0.63, 1.0, 0.35, 0.45, 0.35, 1.0), nrow=3, ncol=3, byrow=TRUE)
m = 1
myfacta(S=S, m=1)
}
