% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T9_3}
\alias{T9_3}
\title{Table 9.3 (Four Measurements of Stiffness)}
\format{
a data frame containing 5 rows and 5 columns:
\itemize{
  \item Taste: The relative quality of the taste of a food
  \item Good buy for money: Value of the food
  \item Flavor: How flavorful the food is
  \item Suitable for snack: Suitability of the food as a snack option
  \item Provides lots of energy: The percieved energy influence of the food
}
}
\source{
Applied Multivariate Statistical Analysis 6th Edition by R. Johnson and D. Wichern
}
\usage{
T9_3
}
\description{
The 5x5 correlation matrix for a data set about food attributes containing 5 variables
}
\keyword{datasets}
